rename test to tests
