+++ src/core/clamp.ts
