+++ src/index.ts
@@ -1,4 +1,4 @@
 export { Timeline } from "./interpol/Timeline"
 export { Interpol } from "./interpol/Interpol"
-export { Ease } from "./helpers/Ease"
+export { Ease } from "~/core/Ease"
 export { psap } from "./psap/psap"
