+++ tests/Interpol.beforeStart.test.ts
