+++ tests/Interpol.pause.test.ts
