+++ tests/Interpol.reverse.test.ts
