+++ tests/Interpol.stop.test.ts
