+++ tests/Timeline.pause.test.ts
