+++ tests/Timeline.stop.test.ts
