+++ tests/psap.specificProperties.test.ts
